Remove scope activation from AWS-SDK instrumentation because for all calls except SQS we aren't expecting to capture any enclosed work. This simplifies the trace handler and means we can turn off various hacks to avoid leaking context when using the async Netty SDK client. (SQS receive calls will be handled separately so we can apply message tracing semantics.)
Instead we activate a no-op scope for the duration of the client call - this protects against enclosed work accidentally keeping the
trace open for longer than necessary. We can also use this to simplify the SQS rework by ignoring when messages are consumed as part
of request handling (for example the AWS client checking MD5 checksums internally) vs user consumption of messages.